+++ components/builder-originsrv/src/data_store.rs
@@ -138,8 +138,8 @@ impl DataStore {
                 &self.vec_to_vec_string(pkg.get_tdeps()),
                 &pkg.get_exposes()
                     .iter()
-                    .map(|e| *e as i16)
-                    .collect::<Vec<i16>>(),
+                    .map(|e| *e as i32)
+                    .collect::<Vec<i32>>(),
                 &PackageVisibility::from(pkg.get_visibility()),
             ],
         ).map_err(SrvError::OriginPackageUpdate)?;
@@ -1499,7 +1499,7 @@ impl DataStore {
         package.set_manifest(row.get("manifest"));
         package.set_config(row.get("config"));
         package.set_target(row.get("target"));
-        let exposes: Vec<i16> = row.get("exposes");
+        let exposes: Vec<i32> = row.get("exposes");
         package.set_exposes(exposes.iter().map(|e| *e as u32).collect::<Vec<u32>>());
         package.set_deps(self.into_idents(row.get("deps")));
         package.set_tdeps(self.into_idents(row.get("tdeps")));
